Added back in the RedditScraper and made an encapsulating class for all scrapers
This saves all information regarding the API to the os system variables with the respective names so Rust can call the values from there instead of needing to call directly from Python. This prevents any unncessary saving of variables and allows the user to not have to set their variables every time using them.